+++ example/App.js
@@ -81,6 +81,7 @@ export default class App extends Component<Props> {
       deviceJSON.hasSystemFeature = ios ? false : await DeviceInfo.hasSystemFeature('amazon.hardware.fire_tv');
       deviceJSON.getSystemAvailableFeatures = ios ? [] : await DeviceInfo.getSystemAvailableFeatures();
       deviceJSON.powerState = ios ? await DeviceInfo.getPowerState() : '';
+      deviceJSON.hasLocationServicesEnabled = await DeviceInfo.hasLocationServicesEnabled();
     } catch (e) {
       console.log('Trouble getting device info ', e);
     }